requirements.txt
pandas==1.0.3
numpy==1.18.1
sklearn==0.0
requests==2.23.0 
PyYAML==5.3.1 